+++ GameEngine/GameEngine/Engine/Core/Engine.cpp
@@ -234,7 +234,7 @@ namespace AGE
 		{
 			return false;
 		}
-		static double refreshStats = 0.0;
+		static float refreshStats = 0.0;
 		refreshStats += _timer->getElapsed();
 		auto &stats = GetThreadManager()->getStatistics();
 		if (refreshStats >= 1.0)
