+++ GameEngine/Tools/AssetsEditor/Scenes/WorldEditorScene.hpp
@@ -15,7 +15,7 @@ namespace AGE
 
 		virtual ~WorldEditorScene(void);
 		virtual bool userStart();
-		virtual bool userUpdateBegin(double time);
-		virtual bool userUpdateEnd(double time);
+		virtual bool userUpdateBegin(float time);
+		virtual bool userUpdateEnd(float time);
 	};
 }
